Extract CodeSnippetUtils to improve testability
